python/projects/py-ego/requirements.txt
pymysql>=1.0.2
pandas>=1.3.0
streamlit>=1.0.0
plotly>=5.0.0
python-dotenv>=1.0.0
openai>=1.0.0
sentence-transformers>=2.2.0